turn Web3AuthProvider into ts format to be built by torus scripts
