+++ packages/composables/modal-vue-composables/package.json
@@ -32,6 +32,7 @@
     "@web3auth/modal": "^8.6.2"
   },
   "devDependencies": {
+    "@toruslabs/openlogin-utils": "^8.2.1",
     "typescript": "^5.2.2"
   },
   "peerDependencies": {
