+++ packages/composables/modal-vue-composables/src/composables/useWeb3Auth.ts
@@ -1,15 +1,10 @@
 import { WalletInitializationError } from "@web3auth/base";
 import { inject } from "vue";
 
-import { IWeb3AuthContext } from "../interfaces";
+import { Web3AuthContextKey } from "../Web3AuthProvider";
 
-export const useWeb3Auth = (): IWeb3AuthContext => {
-  const context = inject<IWeb3AuthContext>("web3AuthContext");
-  if (!context)
-    throw WalletInitializationError.fromCode(
-      1000,
-      // TODO: or is it web3auth context provider
-      "usage of useWeb3Auth not wrapped in `Web3AuthProvider`."
-    );
+export const useWeb3Auth = () => {
+  const context = inject(Web3AuthContextKey);
+  if (!context) throw WalletInitializationError.fromCode(1000, "usage of `useWeb3Auth` not wrapped in `Web3AuthProvider`.");
   return context;
 };
