+++ packages/composables/modal-vue-composables/src/index.ts
@@ -1,3 +1,3 @@
 export * from "./composables";
 export * from "./interfaces";
-export * from "./Web3AuthProvider.vue";
+export { default } from "./Web3AuthProvider";
